requirements.txt
jupyter
Keras==2.4.3
Keras-Preprocessing==1.1.2
matplotlib
numpy==1.18.5
pandas==1.1.5
Pillow==8.3.2
scikit-learn
tensorflow==2.3.4
tqdm==4.62.2